requirements.txt
certifi==2024.8.30
contourpy==1.3.0
cycler==0.12.1
fonttools==4.54.1
geopandas==1.0.1
kiwisolver==1.4.7
matplotlib==3.9.2
meteostat==1.6.8
numpy==2.1.1
packaging==24.1
pandas==2.2.3
pillow==10.4.0
pyogrio==0.10.0
pyparsing==3.1.4
pyproj==3.6.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
shapely==2.0.6
six==1.16.0
tzdata==2024.2
